+++ quill-core/src/main/scala/io/getquill/monad/IOMonad.scala
@@ -38,6 +38,8 @@ trait IOMonad {
 
   object IO {
 
+    def lift[T](result: => Result[T]): IO[T, Effect] = Run(() => result)
+
     def fromTry[T](result: Try[T]): IO[T, Effect] = FromTry(result)
 
     def sequence[A, M[X] <: TraversableOnce[X], E <: Effect](in: M[IO[A, E]])(implicit cbfIOToResult: CanBuildFrom[M[IO[A, E]], Result[A], M[Result[A]]], cbfResultToValue: CanBuildFrom[M[Result[A]], A, M[A]]): IO[M[A], E] =
@@ -80,10 +82,7 @@ trait IOMonad {
       }
 
     def lowerFromTry[U](implicit ev: T => Try[U]) =
-      map(ev).flatMap {
-        case Success(v) => IO.successful(v)
-        case Failure(e) => IO.failed(e)
-      }
+      transform(_.flatMap(ev))
 
     def liftToTry: IO[Try[T], E] =
       transformWith(IO.successful)
@@ -91,15 +90,15 @@ trait IOMonad {
     def failed: IO[Throwable, E] =
       transform {
         case Failure(t) => Success(t)
-        case Success(v) => Failure(new NoSuchElementException("IO.failed not completed with a throwable."))
+        case Success(_) => Failure(new NoSuchElementException("IO.failed not completed with a throwable."))
       }
 
     def map[S](f: T => S): IO[S, E] = transform(_.map(f))
 
     def flatMap[S, E2 <: Effect](f: T => IO[S, E2]): IO[S, E with E2] =
       transformWith {
-        case Success(s) => f(s)
-        case Failure(_) => this.asInstanceOf[IO[S, E with E2]]
+        case Success(s)     => f(s)
+        case f @ Failure(_) => IO.fromTry[S](f.asInstanceOf[Failure[S]]).asInstanceOf[IO[S, E with E2]]
       }
 
     def filter(p: T => Boolean): IO[T, E] =
@@ -117,8 +116,8 @@ trait IOMonad {
 
     def recoverWith[U >: T, E2 <: Effect](pf: PartialFunction[Throwable, IO[U, E2]]): IO[U, E with E2] =
       transformWith {
-        case Failure(t) => pf.applyOrElse(t, (_: Throwable) => this)
-        case Success(_) => this
+        case Failure(t)     => pf.applyOrElse(t, IO.failed _)
+        case s @ Success(_) => IO.fromTry(s)
       }
 
     def zip[S, E2 <: Effect](io: IO[S, E2]): IO[(T, S), E with E2] =