+++ quill-core/src/main/scala/io/getquill/norm/FlattenOptionOperation.scala
@@ -7,10 +7,15 @@ object FlattenOptionOperation extends StatelessTransformer {
   private def isNotEmpty(ast: Ast) =
     BinaryOperation(ast, EqualityOperator.`!=`, NullValue)
 
+  private def emptyOrNot(b: Boolean, ast: Ast) =
+    if (b) OptionIsEmpty(ast) else OptionNonEmpty(ast)
+
   override def apply(ast: Ast): Ast =
     ast match {
       case OptionFlatten(ast) =>
         apply(ast)
+      case OptionGetOrElse(OptionMap(ast, alias, body), Constant(b: Boolean)) =>
+        apply(BinaryOperation(BetaReduction(body, alias -> ast), BooleanOperator.`||`, emptyOrNot(b, ast)): Ast)
       case OptionGetOrElse(ast, body) =>
         apply(If(isNotEmpty(ast), ast, body))
       case OptionFlatMap(ast, alias, body) =>