+++ quill-finagle-postgres/src/main/scala/io/getquill/FinaglePostgresContext.scala
@@ -78,12 +78,12 @@ class FinaglePostgresContext[N <: NamingStrategy](val naming: N, client: Postgre
   def executeBatchAction[B](groups: List[BatchGroup]): Future[List[Long]] = Future.collect {
     groups.map {
       case BatchGroup(sql, prepare) =>
-        prepare.foldLeft(Future.value(List.empty[Long])) {
+        prepare.foldLeft(Future.value(List.newBuilder[Long])) {
           case (acc, prepare) =>
             acc.flatMap { list =>
-              executeAction(sql, prepare).map(list :+ _)
+              executeAction(sql, prepare).map(list += _)
             }
-        }
+        }.map(_.result())
     }
   }.map(_.flatten.toList)
 
@@ -97,12 +97,12 @@ class FinaglePostgresContext[N <: NamingStrategy](val naming: N, client: Postgre
     Future.collect {
       groups.map {
         case BatchGroupReturning(sql, column, prepare) =>
-          prepare.foldLeft(Future.value(List.empty[T])) {
+          prepare.foldLeft(Future.value(List.newBuilder[T])) {
             case (acc, prepare) =>
               acc.flatMap { list =>
-                executeActionReturning(sql, prepare, extractor, column).map(list :+ _)
+                executeActionReturning(sql, prepare, extractor, column).map(list += _)
               }
-          }
+          }.map(_.result())
       }
     }.map(_.flatten.toList)