+++ quill-spark/src/test/scala/io/getquill/context/spark/package.scala
@@ -8,9 +8,13 @@ package object spark {
   val sparkSession =
     SparkSession
       .builder()
-      .master("local")
+      .config("spark.sql.shuffle.partitions", 2) // Default shuffle partitions is 200, too much for tests
+      .config("spark.ui.enabled", "false")
+      .master("local[1]")
       .appName("spark test")
       .getOrCreate()
+
+  sparkSession.sparkContext.setLogLevel("WARN")
 
   implicit val sqlContext = sparkSession.sqlContext